chore: rename file to align with other components
