+++ packages/charts/test/chart-not-animated-styles.js
